move /examples to /tests
- easier to share code
- `cargo test` can run them by default
