+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["examples/*", "sqllogictest", "sqllogictest-bin", "sqllogictest-engines", "tests"]
+members = ["sqllogictest", "sqllogictest-bin", "sqllogictest-engines", "tests"]
 
 [workspace.package]
 version = "0.15.0"
