--- examples/basic/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "basic"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[dependencies]
-sqllogictest = { path = "../../sqllogictest" }
