--- examples/file_level_sort_mode/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "file_level_sort_mode"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[dependencies]
-sqllogictest = { path = "../../sqllogictest" }
