--- examples/test_dir_escape/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "test_dir_escape"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[dependencies]
-sqllogictest = { path = "../../sqllogictest" }
