--- tests/slt/a.slt
@@ -1,4 +1,0 @@
-query T
-hi
-----
-I'm fake.
