--- tests/slt/b/b.slt
@@ -1,4 +1,0 @@
-query T
-hello
-----
-I'm fake.
