+++ tests/slt/basic.slt
