+++ tests/slt/file_level_sort_mode.slt
@@ -1,7 +1,7 @@
 control sortmode rowsort
 
 query III
-select * from example_file_level_sort_mode
+select * from example_sort
 ----
 1 10 2333
 10 100 2333
