+++ tests/slt/rowsort.slt
@@ -1,5 +1,5 @@
 query III rowsort
-select * from example_rowsort
+select * from example_sort
 ----
 1 10 2333
 10 100 2333
