+++ tests/test_dir_escape/test_dir_escape.slt
