+++ tests/validator/validator.rs
@@ -1,5 +1,3 @@
-use std::path::PathBuf;
-
 use sqllogictest::{DBOutput, DefaultColumnType};
 
 pub struct FakeDB;
@@ -27,15 +25,11 @@ impl sqllogictest::DB for FakeDB {
     }
 }
 
-fn main() {
+#[test]
+fn test() {
     let mut tester = sqllogictest::Runner::new(|| async { Ok(FakeDB) });
     // Validator will always return true.
     tester.with_validator(|_, _| true);
 
-    let mut filename = PathBuf::from(file!());
-    filename.pop();
-    filename.pop();
-    filename.push("validator.slt");
-
-    tester.run_file(filename).unwrap();
+    tester.run_file("./validator/validator.slt").unwrap();
 }
